server/requirements.txt
# Core web framework
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.31.1
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
ujson>=5.8.0,<6.0.0
python-multipart>=0.0.6,<0.1.0

# Database
SQLAlchemy[asyncio]>=2.0.18,<3.0.0
alembic>=1.11.1,<2.0.0
asyncpg>=0.30.0

# Cache
redis[hiredis]>=4.6.0,<5.0.0

# Logging
loguru>=0.7.0,<0.8.0

# Auth (JWT)
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==4.0.1
python-jose>=3.3.0,<4.0.0

# Utilities
yarl>=1.9.2,<2.0.0
aiofiles>=23.1.0,<24.0.0
httptools>=0.6.0,<0.7.0

# HTTP 客户端 (调用 Agent Service)
httpx>=0.28.1

# FastMCP 客户端
fastmcp~=2.11.1